+++ stdlib/LinearAlgebra/test/cholesky.jl
@@ -22,7 +22,7 @@ function unary_ops_tests(a, ca, tol; n=size(a, 1))
     @test logabsdet_ca[1] ≈ logabsdet_a[1]
     @test logabsdet_ca[2] ≈ logabsdet_a[2]
     @test isposdef(ca)
-    @test_throws ErrorException ca.Z
+    @test_throws FieldError ca.Z
     @test size(ca) == size(a)
     @test Array(copy(ca)) ≈ a
     @test tr(ca) ≈ tr(a) skip=ca isa CholeskyPivoted
@@ -72,9 +72,14 @@ end
         ε = εa = eps(abs(float(one(eltya))))
 
         # Test of symmetric pos. def. strided matrix
+<<<<<<< dk/cholfix
         apd  = Matrix(Hermitian(a'*a))
         @inferred cholesky(apd)
         capd  = factorize(apd)
+=======
+        apd  = a'*a
+        capd  = @inferred cholesky(apd)
+>>>>>>> master
         r     = capd.U
         κ     = cond(apd, 1) #condition number
 
@@ -82,7 +87,7 @@ end
         if eltya != Int
             @test Factorization{eltya}(capd) === capd
             if eltya <: Real
-                @test Array(Factorization{complex(eltya)}(capd)) ≈ Array(factorize(complex(apd)))
+                @test Array(Factorization{complex(eltya)}(capd)) ≈ Array(cholesky(complex(apd)))
                 @test eltype(Factorization{complex(eltya)}(capd)) == complex(eltya)
             end
         end
@@ -392,7 +397,7 @@ end
         0.11192755545114 - 0.1603741874112385im 0.8439562576196216 + 1.0850814110398734im
         -1.0568488936791578 - 0.06025820467086475im 0.12696236014017806 - 0.09853584666755086im]
     cholesky(Hermitian(apd, :L), RowMaximum()) \ b
-    r = factorize(apd).U
+    r = cholesky(apd).U
     E = abs.(apd - r'*r)
     ε = eps(abs(float(one(ComplexF32))))
     n = 10